# documentation/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service to store a high-integrity event register.",
        "version": "1.0.0",
        "title": "Register"
    },
    "host": "localhost:8005",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/entry/{entry_number}": {
            "get": {
                "description": "Get the specified entry (register update event)",
                "operationId": "getEntry",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "entry_number",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entry"
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/record": {
            "post": {
                "description": "Add a new event record to the register. See Validation API for possible errors.",
                "operationId": "postRecord",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Item data to be added to the register",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/minted-item"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "entry_number": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/record/{field_value}": {
            "get": {
                "description": "Retrieve the most up to date infoset for the record with primary key \"field_value\"",
                "operationId": "getRecordFieldValue",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "field_value",
                        "description": "The primary key of the record to retrieve",
                        "required": true,
                        "type": "string",
                        "format": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/infoset"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/record/{field_value}/entries": {
            "get": {
                "description": "Retrieve all entries with primary key \"field_value\"",
                "operationId": "getRecordFieldValueEntries",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "field_value",
                        "description": "The record's primary key",
                        "required": true,
                        "type": "string",
                        "format": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entry"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/records": {
            "get": {
                "description": "Get all records in the register",
                "operationId": "getRecords",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/records"
                        }
                    }
                }
            },
            "post": {
                "description": "Create multiple new records",
                "operationId": "postRecords",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Item data to be added to the register",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/minted-item"
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "entry_number": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/records/{field_name}/{field_value}": {
            "get": {
                "description": "Get all records with the specified value in the specified field",
                "operationId": "getRecordsNameValue",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "field_name",
                        "description": "The name of the field to match against",
                        "required": true,
                        "type": "string",
                        "format": ""
                    },
                    {
                        "in": "path",
                        "name": "field_value",
                        "description": "The value of the field to query on",
                        "required": true,
                        "type": "string",
                        "format": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/records"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entry": {
            "type": "object",
            "properties": {
                "entry_number": {
                    "type": "integer"
                },
                "entry_timestamp": {
                    "type": "string"
                },
                "item_hash": {
                    "type": "string"
                },
                "key": {
                    "description": "Primary identifier for the register",
                    "type": "string"
                },
                "item_signature": {
                    "type": "string"
                }
            }
        },
        "minted-item": {
            "type": "object",
            "properties": {
                "item": {
                    "type": "object",
                    "description": "The schema of the item property is undefined for the register API, but is validated by validation-api, if specified."
                },
                "item-hash": {
                    "type": "string"
                },
                "item-signature": {
                    "type": "string"
                }
            }
        },
        "infoset": {
            "type": "object",
            "properties": {
                "entry_number": {
                    "type": "integer"
                },
                "entry_timestamp": {
                    "type": "string"
                },
                "item_hash": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "item": {
                    "type": "object",
                    "description": "The schema of the item is variable and enforced by Validation API"
                }
            }
        },
        "records": {
            "type": "object",
            "description": "Note: contains object for each infoset in the register.",
            "properties": {
                "<primary_key>": {
                    "$ref": "#/definitions/infoset"
                },
                "<primary_key_2>": {
                    "$ref": "#/definitions/infoset"
                }
            }
        },
...
    }
}